Added referer data to custom variable 3.
Wanted to do this directly through Google Analytics' reporting mechanism
since this is one of variables that they support directly, but it seems
that this is not working with this version of the GA API we're using. We
spiked a bit on using the JS ga script and we confirmed that we could
set the Referer from there.

The most likely scenario is that Google broke this older API at some
point because most clients are really just using the JS library they
provide. We could spike further on making a ruby gem for the new API but
that would take some time.

[Finishes #73346456]