Pass HTTP_ACCEPT_LANGUAGE directly
The previous code would take the first thing that matched a pattern, \w\w-\w\w, and try to normalize it to e.g. en-US (from en-us).

The W3 spec for the Accept-Language header describes the ability to pass multiple languages, with a weight for each: http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.4.

For now we will just pass along whatever the Accept-Language is unless it's for some reason a problem, and make the appropriate change at that time.